+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Added HTTP Access Log middleware
     - Added command methods to RPC Client
     - Removed Client interface
+    - Added Bzip2 message serializer
 
 - 0.2.1-beta
     - Support PSR-15 Server Request Handler interface